improve reusability of icon tooltip, add tooltip to summary title
