+++ context/app/static/js/shared-styles/icons/TooltipIcon.tsx
@@ -1,0 +1,24 @@
+import React from 'react';
+import { styled } from '@mui/material/styles';
+import { InfoIcon } from 'js/shared-styles/icons';
+import { SecondaryBackgroundTooltip } from '../tooltips';
+
+const StyledInfoIcon = styled(InfoIcon)(({ theme }) => ({
+  marginLeft: theme.spacing(0.5),
+  fontSize: '1.5rem',
+}));
+
+interface TooltipIconProps {
+  iconTooltipText?: string;
+}
+
+export default function InfoTooltipIcon({ iconTooltipText }: TooltipIconProps) {
+  if (!iconTooltipText) {
+    return null;
+  }
+  return (
+    <SecondaryBackgroundTooltip title={iconTooltipText}>
+      <StyledInfoIcon color="primary" />
+    </SecondaryBackgroundTooltip>
+  );
+}
